requirements.txt
altgraph==0.17.5
bcrypt==5.0.0
click==8.3.1
dlib==20.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
macholib==1.16.4
numpy==2.2.6
opencv-python==4.12.0.88
packaging==26.0
pillow==12.0.0
psycopg2-binary==2.9.11
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.1
PyQt6==6.10.0
PyQt6-Qt6==6.10.0
PyQt6_sip==13.10.2
scipy==1.16.3
setuptools==80.9.0
